requirements.txt
click==8.1.7
rich==13.7.1
PyYAML==6.0.1
jsonschema==4.23.0
python-dotenv==1.0.1
